requirements.txt
# Deep Learning & AI
tensorflow>=2.11.0
keras>=2.11.0
numpy>=1.23.0

# Web Framework
Flask>=2.2.0
Flask-Session>=0.4.0
Werkzeug>=2.2.0

# Database
pymongo>=4.3.0
python-dotenv>=0.20.0

# Image Processing
opencv-python>=4.6.0
Pillow>=9.0.0
python-magic>=0.4.24

# Security & Validation
python-dotenv>=0.20.0

# Testing & Documentation
pytest>=7.0.0
pytest-cov>=4.0.0
flasgger>=0.9.0

# Additional Utilities
scipy>=1.9.0
matplotlib>=3.6.0
requests>=2.28.0

# PDF Generation (optional)
reportlab>=4.0.0
pypdf2>=3.0.0
